I renamed the m3d notebook to something more descriptive
I added the tutorial for the medvae, and it requires a local pip installation of the medvae repo.

I didn't use a submodule for simplicity, although this may need to be revised later. If so, I need to add a git clone line to the readme in the medvae folder.